refactor(end-to-end): Move teardown method to superclass
